# store/copr: remove useless ticker for performance